Desktop Monday night

+++ phd_journal/24_12_03/plotting.py
@@ -1,5 +1,5 @@
 import itertools
-from os.path import join
+from os.path import join, exists
 from pickle import load
 
 import matplotlib.pyplot as plt
@@ -13,6 +13,7 @@ from scikitplot.metrics import plot_roc_curve
 from scipy.stats import shapiro
 from sklearn.metrics import classification_report, confusion_matrix, roc_curve, roc_auc_score
 import neptune.integrations.sklearn as npt_utils
+import plotly.express as px
 
 
 image_format = "png"
@@ -224,14 +225,19 @@ def plot_classification_with_var(run, in_path, out_path, var_of_interest, varian
     pred = classification_res['pred']
     true = classification_res['true']
 
-    # Make sample weights (n_samples, ) according to the class size
-    classes = true.unique()
-    class_weight = {c: len(true) / (len(classes) * len(true[true == c])) for c in classes}
-    sample_weight = true.map(class_weight)
-
     # Compute classification metrics
-    report = classification_report(true, pred, output_dict=True, sample_weight=sample_weight)
+    report = classification_report(true, pred, output_dict=True)
     run[f"classification/{var_of_interest}/my_metrics"] = report
+
+    # Compute classification metrics by dataset
+    # The indexes can be separated by '-' to get the dataset name
+    datasets = set([i.split("-")[0] for i in true.index])
+    for dataset in datasets:
+        idx = [i for i in true.index if i.startswith(dataset)]
+        report_dataset = classification_report(true.loc[idx], pred.loc[idx], output_dict=True)
+        run[f"classification/{var_of_interest}/{dataset}"] = report_dataset
+
+    print('W. F1-Score:', report['weighted avg']['f1-score'])
 
     # Plot weighted F1-Score
     fig = plt.figure(figsize=(3, 4))
@@ -264,30 +270,32 @@ def plot_classification_with_var(run, in_path, out_path, var_of_interest, varian
     """
 
 def plot_2components(run, _metadata, in_path, out_path, method):
-    # Read csv principal components file
-    pc = pd.read_csv(join(in_path, f"{method}_transformed.csv"), index_col=0)
+    filepath = join(in_path, f"{method}_transformed.csv")
+    if exists(filepath):
+        # Read csv principal components file
+        pc = pd.read_csv(filepath, index_col=0)
 
-    # Plot stlyles
-    dataset_colors = {"Newcastle": "blue", "Izmir": "green", "Istambul": "orange", "Miltiadous": "pink", "BrainLat:AR": "black", "BrainLat:CL": "yellow"}
-    #dataset_colors = {l: [c / 255 for c in color] for l, color in dataset_colors.items()}
-    diagnoses_circles = {"HC": "x", "AD": "o"}
+        # Plot stlyles
+        dataset_colors = {"Newcastle": "blue", "Izmir": "green", "Istambul": "orange", "Miltiadous": "pink", "BrainLat:AR": "black", "BrainLat:CL": "yellow"}
+        #dataset_colors = {l: [c / 255 for c in color] for l, color in dataset_colors.items()}
+        diagnoses_circles = {"HC": "x", "AD": "o"}
 
-    # One plot, all datasets
-    fig = plt.figure(figsize=(6, 6))
-    for dataset, color in dataset_colors.items():
-        for diagnosis, marker in diagnoses_circles.items():
-            idx = _metadata[(_metadata["DIAGNOSIS"] == diagnosis) & (_metadata["SITE"] == dataset)].index
-            existing_idx = pc.index.intersection(idx)
-            plt.scatter(pc['0'].loc[existing_idx], pc['1'].loc[existing_idx], color=color,
-                        label=f"{dataset} - {diagnosis}", marker=marker)
+        # One plot, all datasets
+        fig = plt.figure(figsize=(6, 6))
+        for dataset, color in dataset_colors.items():
+            for diagnosis, marker in diagnoses_circles.items():
+                idx = _metadata[(_metadata["DIAGNOSIS"] == diagnosis) & (_metadata["SITE"] == dataset)].index
+                existing_idx = pc.index.intersection(idx)
+                plt.scatter(pc['0'].loc[existing_idx], pc['1'].loc[existing_idx], color=color,
+                            label=f"{dataset} - {diagnosis}", marker=marker)
 
-    plt.legend(loc='best', fontsize=8)
-    plt.xlabel("Component 1", fontsize=14)
-    plt.ylabel("Component 2", fontsize=14)
-    plt.tight_layout()
-    run[f"quality_control/{method}"].upload(File.as_image(plt.gcf()))
-    plt.savefig(join(out_path, f"{method}.{image_format}"), dpi=dpi, transparent=True, bbox_inches='tight')
-    plt.close(fig)
+        plt.legend(loc='best', fontsize=8)
+        plt.xlabel("Component 1", fontsize=14)
+        plt.ylabel("Component 2", fontsize=14)
+        plt.tight_layout()
+        run[f"quality_control/{method}"].upload(File.as_image(plt.gcf()))
+        plt.savefig(join(out_path, f"{method}.{image_format}"), dpi=dpi, transparent=True, bbox_inches='tight')
+        plt.close(fig)
 
 def plot_distance_matrix(run, in_path, out_path):
     # Distance matrices show the Euclidean distance across all features between batch/site-average values
@@ -413,24 +421,53 @@ def plot_simple_diagnosis_discriminant(run, datasets_metadata, in_path, out_path
     model = load(open(join(in_path, "simple_diagnosis_discriminant.pkl"), 'rb'))
 
     # Plot stlyles
-    dataset_colors = {"Newcastle": "blue", "Izmir": "green", "Istambul": "orange", "Miltiadous": "pink", "BrainLat:AR": "black", "BrainLat:CL": "yellow"}
-    #dataset_colors = {l: [c / 255 for c in color] for l, color in dataset_colors.items()}
-    diagnoses_circles = {"HC": "x", "AD": "o"}
+    diagnoses_colors = {"HC": "blue", "AD": "red"}
 
     # One plot, all datasets
-    fig = plt.figure(figsize=(6, 6))
-    for dataset, color in dataset_colors.items():
-        for diagnosis, marker in diagnoses_circles.items():
-            idx = datasets_metadata[(datasets_metadata["DIAGNOSIS"] == diagnosis) & (datasets_metadata["SITE"] == dataset)].index
+    if len(pc.columns) <= 2:
+        fig = plt.figure(figsize=(6, 6))
+        for diagnosis, color in diagnoses_colors.items():
+            idx = datasets_metadata[datasets_metadata["DIAGNOSIS"] == diagnosis].index
             existing_idx = pc.index.intersection(idx)
-            # plot 1d
-            plt.scatter(pc['0'].loc[existing_idx], np.zeros(len(existing_idx)), color=color, label=f"{dataset} - {diagnosis}", marker=marker)
+            match len(pc.columns):
+                case 1:
+                    plt.scatter(pc['0'].loc[existing_idx], np.zeros(len(existing_idx)), color=color, label=diagnosis, alpha=0.5)
+                    plt.xlabel("Component 1", fontsize=14)
+                case 2:
+                    plt.scatter(pc['0'].loc[existing_idx], pc['1'].loc[existing_idx], color=color, label=diagnosis, alpha=0.5)
+                    plt.xlabel("Component 1", fontsize=14)
+                    plt.ylabel("Component 2", fontsize=14)
+    else:
+        # Number of pair-wise combinations?
+        num_combinations = len(list(itertools.combinations(pc.columns, 2)))
+        num_cols = 3
+        num_rows = (num_combinations + num_cols - 1) // num_cols
+        fig = plt.figure(figsize=(4 * num_cols, 3 * num_rows))
+        for i, (col1, col2) in enumerate(itertools.combinations(pc.columns, 2)):
+            plt.subplot(num_rows, num_cols, i + 1)
+            for diagnosis, color in diagnoses_colors.items():
+                idx = datasets_metadata[datasets_metadata["DIAGNOSIS"] == diagnosis].index
+                existing_idx = pc.index.intersection(idx)
+                plt.scatter(pc[col1].loc[existing_idx], pc[col2].loc[existing_idx], color=color, label=diagnosis, alpha=0.5)
+                plt.xlabel(f"Component {col1}", fontsize=14)
+                plt.ylabel(f"Component {col2}", fontsize=14)
+
+            """
+            ax = fig.add_subplot(projection='3d')
+            ax.scatter(
+                pc.loc[existing_idx, '0'],
+                pc.loc[existing_idx, '1'],
+                pc.loc[existing_idx, '2'],
+                color=color, label=diagnosis
+            )
+            ax.set_xlabel('Component 1')
+            ax.set_ylabel('Component 2')
+            ax.set_zlabel('Component 3')
+            plt.show(block=True)
+            """
 
     plt.legend(loc='best', fontsize=8)
-    plt.xlabel("Component 1", fontsize=14)
-    plt.ylabel("Component 2", fontsize=14)
     plt.tight_layout()
     run[f"simple_diagnosis_discriminant/visual_separation"].upload(File.as_image(plt.gcf()))
     plt.savefig(join(out_path, f"simple_diagnosis_discriminant.{image_format}"), dpi=dpi, transparent=True, bbox_inches='tight')
     plt.close(fig)
-
